a little change
Please refer to UpdateBrainValue function and RayDetect Update function

+++ Assets/Scripts/DeviceManager.cs
@@ -28,7 +28,7 @@ public class DeviceManager : MonoBehaviour {
             gazePoint = lastGazePoint.Screen;    
         }
 	}
-    public float GetBrainValue()
+    public int GetBrainValue()
     {
         return brainValue;
     }
